refactor(pdf-overlay): extract presentation and render-cache flow
