+++ src/app/overlays/pdf/present.rs
@@ -1,0 +1,174 @@
+use super::DisplayedPdfPreview;
+use crate::app::App;
+use crate::app::overlays::inline_image::{
+    ImageProtocol, OverlayPresentState, place_terminal_image, preview_log,
+};
+use anyhow::{Context, Result};
+use ratatui::layout::Rect;
+
+impl App {
+    pub(in crate::app) fn present_pdf_overlay(
+        &mut self,
+        protocol: ImageProtocol,
+        excluded: &[Rect],
+        out: &mut Vec<u8>,
+    ) -> Result<OverlayPresentState> {
+        let Some(request) = self.active_pdf_overlay_request() else {
+            preview_log("present_pdf_overlay: no request");
+            return Ok(OverlayPresentState::NotRequested);
+        };
+        preview_log(format_args!(
+            "present_pdf_overlay: path={:?} page={}",
+            request.path, request.page
+        ));
+
+        if !self.pdf_selection_activation_ready() {
+            preview_log("present_pdf_overlay: activation not ready → Waiting");
+            return Ok(OverlayPresentState::Waiting);
+        }
+
+        let Some(requested_placement) = self.overlay_placement_for_request(&request) else {
+            preview_log("present_pdf_overlay: no placement yet → probe + Waiting");
+            let _ = self.ensure_pdf_page_probe(&request);
+            return Ok(OverlayPresentState::Waiting);
+        };
+        let render_key = self.pdf_render_key_from_request(&request, requested_placement);
+        let Some(rendered) = self.ensure_pdf_render(&render_key) else {
+            preview_log("present_pdf_overlay: render not ready → Waiting");
+            return Ok(OverlayPresentState::Waiting);
+        };
+        let placement = self.resolved_pdf_display_placement(
+            &request,
+            &render_key,
+            requested_placement,
+            &rendered,
+        );
+        preview_log(format_args!(
+            "present_pdf_overlay: placement={:?}",
+            placement.image_area
+        ));
+        let displayed = DisplayedPdfPreview::from_request(&request, placement);
+        let image_changed = self.pdf_preview.displayed.as_ref() != Some(&displayed);
+        let excluded_changed = excluded != self.pdf_preview.displayed_excluded.as_slice();
+        if !image_changed && !excluded_changed {
+            preview_log("present_pdf_overlay: already displayed → Displayed");
+            return Ok(OverlayPresentState::Displayed);
+        }
+        if image_changed {
+            out.extend(self.clear_preview_overlay()?);
+        }
+        let bytes = place_terminal_image(protocol, &rendered, placement.image_area, excluded, None)
+            .context("failed to display PDF page")?;
+        preview_log(format_args!(
+            "present_pdf_overlay: placed {} bytes via {protocol:?}",
+            bytes.len()
+        ));
+        out.extend(bytes);
+        self.pdf_preview.displayed = Some(displayed);
+        self.pdf_preview.displayed_excluded = excluded.to_vec();
+        Ok(OverlayPresentState::Displayed)
+    }
+
+    pub(crate) fn preview_prefers_pdf_surface(&self) -> bool {
+        if !self.terminal_image_overlay_available()
+            || !self.pdf_preview.pdf_tools_available
+            || self.pdf_preview.session.is_none()
+        {
+            return false;
+        }
+        if self.preview_uses_image_overlay() {
+            return true;
+        }
+
+        let Some(request) = self.active_pdf_overlay_request() else {
+            return false;
+        };
+        if !self.pdf_selection_activation_ready() {
+            return true;
+        }
+
+        let page_key = self.pdf_page_key_from_request(&request);
+        if self.pdf_preview.failed_page_probes.contains(&page_key) {
+            return false;
+        }
+        if self.pdf_preview.pending_page_probes.contains(&page_key)
+            || !self.pdf_preview.page_dimensions.contains_key(&page_key)
+        {
+            return true;
+        }
+
+        let Some(placement) = self.overlay_placement_for_request(&request) else {
+            return false;
+        };
+        let render_key = self.pdf_render_key_from_request(&request, placement);
+        if self.pdf_preview.failed_renders.contains(&render_key) {
+            return false;
+        }
+        self.pdf_preview.pending_renders.contains(&render_key)
+            || self.cached_render_exists(&render_key)
+    }
+
+    pub(crate) fn preview_overlay_placeholder_message(&self) -> Option<String> {
+        if self.preview_prefers_static_image_surface() && !self.preview_uses_image_overlay() {
+            return self.static_image_overlay_placeholder_message();
+        }
+
+        if !self.preview_prefers_pdf_surface() || self.preview_uses_image_overlay() {
+            return None;
+        }
+
+        let request = self.active_pdf_overlay_request()?;
+        let page_key = self.pdf_page_key_from_request(&request);
+        if self.pdf_preview.failed_page_probes.contains(&page_key) {
+            return Some("PDF preview unavailable".to_string());
+        }
+        if !self.pdf_selection_activation_ready()
+            || !self.pdf_preview.page_dimensions.contains_key(&page_key)
+            || self.pdf_preview.pending_page_probes.contains(&page_key)
+        {
+            return None;
+        }
+
+        let placement = self.overlay_placement_for_request(&request)?;
+        let render_key = self.pdf_render_key_from_request(&request, placement);
+        if self.pdf_preview.failed_renders.contains(&render_key) {
+            return Some("PDF preview unavailable".to_string());
+        }
+        if self.cached_render_exists(&render_key) {
+            return None;
+        }
+        None
+    }
+
+    pub(in crate::app) fn pdf_overlay_displayed(&self) -> bool {
+        self.pdf_preview.displayed.is_some()
+    }
+
+    pub(in crate::app) fn displayed_pdf_overlay_area(&self) -> Option<Rect> {
+        self.pdf_preview
+            .displayed
+            .as_ref()
+            .map(|displayed| displayed.area)
+    }
+
+    pub(in crate::app) fn clear_displayed_pdf_overlay(&mut self) {
+        self.pdf_preview.displayed = None;
+    }
+
+    pub(in crate::app) fn displayed_pdf_overlay_matches_active(&self) -> bool {
+        self.active_pdf_display_target()
+            .as_ref()
+            .zip(self.pdf_preview.displayed.as_ref())
+            .is_some_and(|(active, displayed)| active == displayed)
+    }
+
+    fn active_pdf_display_target(&self) -> Option<DisplayedPdfPreview> {
+        let request = self.active_pdf_overlay_request()?;
+        if !self.pdf_selection_activation_ready() {
+            return None;
+        }
+        let requested_placement = self.overlay_placement_for_request(&request)?;
+        let placement = self.cached_display_placement_for_request(&request, requested_placement)?;
+        Some(DisplayedPdfPreview::from_request(&request, placement))
+    }
+}
